+++ core/views/__init__.py
@@ -10,5 +10,5 @@ from .field_mapper import *
 from .published import *
 from .record_identifer_transformation_scenario import *
 from .stateio import *
-from .background_tasks import *
+from .core_background_tasks import *
 from .export import *
